Fix Auth Again

+++ src/main/java/org/cnc/cncbot/dto/publicplayerinfo/C.java
@@ -12,7 +12,7 @@ private Integer i;
 
 private String n;
 
-private Integer p;
+private Long p;
 
 private Integer x;
